update gRPC connection for exercise-service
